+++ src/main/java/com/nylas/ComponentQuery.java
@@ -1,0 +1,4 @@
+package com.nylas;
+
+public class ComponentQuery extends RestfulQuery<ComponentQuery> {
+}