+++ src/main/java/com/nylas/JsonHelper.java
@@ -2,14 +2,18 @@ package com.nylas;
 
 import java.io.IOException;
 import java.lang.reflect.Type;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import com.nylas.delta.Delta;
 import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
 
+import com.squareup.moshi.adapters.Rfc3339DateJsonAdapter;
 import okhttp3.MediaType;
 import okhttp3.RequestBody;
 
@@ -21,7 +25,9 @@ public class JsonHelper {
 		moshi = new Moshi.Builder()
 				.add(Event.WHEN_JSON_FACTORY)
 				.add(Event.EVENT_NOTIFICATION_JSON_FACTORY)
-				.add(Delta.ACCOUNT_OWNED_MODEL_JSON_FACTORY)
+        .add(Delta.ACCOUNT_OWNED_MODEL_JSON_FACTORY)
+				.add(new NeuralCategorizer.CategorizeCustomAdapter())
+				.add(Date.class, new Rfc3339DateJsonAdapter().nullSafe())
 				.build();
 	}
 	
@@ -55,6 +61,17 @@ public class JsonHelper {
 	
 	public static Map<String, Object> jsonToMap(String json) {
 		return fromJsonUnchecked(mapAdapter, json);
+	}
+
+	public static Map<String, Object> jsonToMap(JsonReader jsonReader) throws IOException {
+		Map<String, Object> json = new HashMap<>();
+		jsonReader.beginObject();
+		while(jsonReader.hasNext()) {
+			json.put(jsonReader.nextName(), jsonReader.readJsonValue());
+		}
+		jsonReader.endObject();
+
+		return json;
 	}
 	
 	private static final MediaType jsonType = MediaType.parse("application/json; charset=utf-8");