+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+- Added Outbox support
+- Added support for Component CRUD
+- Added support for Neural Categorizer
 - Added support for Delta
 
 ### Changed